refactor: collapse three thin single-purpose files into their natural homes
